+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.7.3 (????-??-??)
 
 - Calling `Pagerfanta::setMaxPerPage()` should not reset the number of results
+- Avoid using deprecated `Doctrine\DBAL\Query\QueryBuilder::execute()` when able
 
 ## 2.7.2 (2021-08-01)